Group project code submitted as faculty's advice
This was a group task. In the presentation, we presented an existing paper, published by some extra-ordinary south-indian authors that was focused on Transfer learning‑based deep ensemble neural network for plant leaf disease detection.

After the presentation the faculty assigned us a project, where our task was to find the limitations and the future scopes of the research that we presented. We tried hard to find the original codes from the authors, but we were unable to find the codes. Therefore, we tried ourselves to work on the scopes and the limitations and we used some existing models to work on the limitations and the scopes that we could work on.  So this is the result of our group work.

Submitted to: DR. DEBAJYOTI KARMAKER, Faculty, AIUB.

Submitted by:
CVPR[B] Group-10
MD. ASIFUR RAHMAN -[20-43064-1]
ABRAR IBNE AHSAN - -[19-41389-3]
SHINZON SIDDIQUA - -[20-43671-2]
H.M.Saifullahil Mazid- -[20-44018-2]